+++ AmoRepository/Models/Contact.cs
@@ -47,11 +47,11 @@ namespace MZPO.AmoRepo
         /// <summary>
         /// Дата создания контакта, передается в Unix Timestamp.
         /// </summary>
-        public int? created_at { get; set; }
+        public long? created_at { get; set; }
         /// <summary>
         /// Дата изменения контакта, передается в Unix Timestamp.
         /// </summary>
-        public int? updated_at { get; set; }
+        public long? updated_at { get; set; }
         /// <summary>
         /// Дата ближайшей задачи к выполнению, передается в Unix Timestamp.
         /// </summary>
@@ -73,41 +73,6 @@ namespace MZPO.AmoRepo
         /// </summary>
         public Embedded _embedded { get; set; }
 
-        //public class Custom_fields_value
-        //{
-        //    /// <summary>
-        //    /// ID поля.
-        //    /// </summary>
-        //    public int field_id { get; set; }
-        //    /// <summary>
-        //    /// Название поля.
-        //    /// </summary>
-        //    public string field_name { get; set; }
-        //    /// <summary>
-        //    /// Код поля.
-        //    /// </summary>
-        //    public string field_code { get; set; }
-        //    /// <summary>
-        //    /// Тип поля.
-        //    /// </summary>
-        //    public string field_type { get; set; }
-        //    /// <summary>
-        //    /// Массив значений поля
-        //    /// </summary>
-        //    public Values[] values { get; set; }
-
-        //    public class Values
-        //    {
-        //        /// <summary>
-        //        /// Значения поля.
-        //        /// </summary>
-        //        [JsonProperty(NullValueHandling = NullValueHandling.Include)]
-        //        public object value { get; set; }
-        //        public int? enum_id { get; set; }
-        //        public string enum_code { get; set; }
-        //    }
-        //}
-        
         public class Links
         {
             public Link self { get; set; }
