+++ ReportProcessors/Processors/RetailKPIProcessor.cs
@@ -343,7 +343,7 @@ namespace MZPO.ReportProcessors
             if (finishedLeads > 0)
                 averageTime = allLeads.AsParallel()
                     .Where(x => (x.status_id == 142) || (x.status_id == 143))
-                    .Select(x => (int)x.closed_at - (int)x.created_at).Average() / 86400;
+                    .Select(x => (long)x.closed_at - (long)x.created_at).Average() / 86400;
 
             List<Request> requestContainer = new();
             requestContainer.Add(GetRowRequest(manager.Item1, GetCellData(dates, totalNewLeads, finishedLeads, successLeads, totalSales, averageTime, responseTime, longLeads, inCallsCount, outCallsCount, missedCallsCount)));
